separate parameter data fixtures data class into subdirectory
